Adding top-level fields to CLI commands
The top-level fields are currently not accessible from the CLI, which
makes it impossible to select entries or groups based on the UUID.
There are other top-level fields I believe, like the expiry date, but
I only added the two most critical fields for now. I also considered
defining the constants in `Entry` instead of `EntryAttributes`, because
they are not attributes per se. I'm not sure which choice is better tbh.